# backend/openapi.yaml
openapi: 3.1.0
info:
  title: PodChat Backend API
  version: 1.0.0
  description: |
    REST API for PodChat — live podcast platform with rooms, recordings,
    LiveKit audio and Socket.IO realtime presence.
servers:
  - url: http://localhost:3001/api
    description: Local dev
  - url: https://livepodchat.com/api
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Recording:
      type: object
      properties:
        id: { type: string }
        roomId: { type: string }
        fileUrl: { type: string }
        durationSeconds: { type: integer }
        fileSizeBytes: { type: integer }
        format: { type: string }
        createdAt: { type: string, format: date-time }
        title: { type: string }
        description: { type: string }
        shareSlug: { type: string }
        isPublic: { type: boolean }
        playCount: { type: integer }
...
security:
  - bearerAuth: []
paths:
...
  /rooms/{slug}/recordings:
    get:
      tags: [recordings]
      summary: Recordings for a room
      parameters:
        - { in: path, name: slug, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Recording' }
  /recordings/{id}/download:
    get:
      tags: [recordings]
      summary: Get a presigned download URL for a recording
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
                  expiresAt: { type: string, format: date-time }
...